Huge improvements to stage 2 of P-1 when lots of memory is available. Warning: upgrading in the middle of P-1 stage 2 will restart P-1 stage 2 from scratch

+++ pm1prob.h
@@ -11,10 +11,10 @@ extern "C" {
 #endif
 
 // Returns the probability of PM1(B1,B2) success for a Mersenne 2^exponent -1 already TF'ed to factoredUpTo.
-double pm1prob(double takeAwayBits, unsigned factoredUpTo, double B1, double B2);
+double pm1prob (double takeAwayBits, unsigned factoredUpTo, uint64_t B1, uint64_t B2);
 
 // Dickman's "rho" function; rho(x) == F(1/x)
-double rho(double x);
+double rho (double x);
 
 #ifdef __cplusplus
 }
